Add watchfiles attribute to contents manager
