+++ jupyter_server/services/contents/manager.py
@@ -34,6 +34,13 @@ from .checkpoints import AsyncCheckpoints, Checkpoints
 copy_pat = re.compile(r"\-Copy\d*\.")
 
 
+class NoWatchfilesAPI:
+    pass
+
+
+NOWATCHFILESAPI = NoWatchfilesAPI()
+
+
 class ContentsManager(LoggingConfigurable):
     """Base class for serving files and directories.
 
@@ -408,6 +415,20 @@ class ContentsManager(LoggingConfigurable):
 
     # ContentsManager API part 2: methods that have useable default
     # implementations, but can be overridden in subclasses.
+
+    @property
+    def watchfiles(self):
+        """File system change notifyer
+
+        Override this method in subclasses if the file system supports change notifications.
+
+        Returns
+        -------
+        api : class
+            The supported API for file system change notifications. Loosely follows the API of the
+            watchfiles Python package (can be a subset of it).
+        """
+        return NOWATCHFILESAPI
 
     def delete(self, path):
         """Delete a file/directory and any associated checkpoints."""
