+++ setup.cfg
@@ -53,7 +53,7 @@ install_requires =
     terminado>=0.8.3
     tornado>=6.1.0
     traitlets>=5.1
-    watchfiles>=0.12
+    watchfiles>=0.13
     websocket-client
 
 [options.extras_require]
